** NOTE ** The code at this point is really ugly due to all debugging and changing and rewriting stuff. Please allow for some cleaning up before checking through it.
Made knownFiles to contain just pointers.
Added flushBitmask() function to write the bitmask to a .bmask file.
Added progressFromBitmask() which computes the progress in percentages from a given KnownFile, using its number of packages, and the number of packets that are downloaded.
Made FbpClient temporarily start a file download right away, instead of waiting for user input for it.
Made FbpClient send out more warnings when servers or ID's overlap
Made FbpClient send requests when the server is waiting for them
Fixed some bugs in FbpClient, for example regarding packets smaller than the normal packet size; this is legal for the last packet, and that it would take knownFiles_[id] erroneously because id != index in knownFiles_
Write data correctly, with correct dataSize and flushing afterwards
Made the Fbp UI display file progress too.
Made dataFile open in ReadWrite mode, so it isn't truncated on open() as with WriteOnly